+++ package/Source~/build_android.sh
@@ -1,8 +1,8 @@
 #!/bin/sh
 
 if [ "$ANDROID_NDK_ROOT" == "" ]; then
-	echo "ERROR: Path to Android NDK is not setup. Please make sure you set the ANDROID_NDK env variable to the location of your NDK installation!"
-	exit -1
+    echo "ERROR: Path to Android NDK is not setup. Please make sure you set the ANDROID_NDK env variable to the location of your NDK installation!"
+    exit -1
 fi
 
 PROJECT=UnityARCore
@@ -12,7 +12,7 @@ $ANDROID_NDK_ROOT/build/ndk-build
 TARGET_FOLDER=../Runtime/Android
 
 if [ ! -e "$TARGET_FOLDER" ]; then
-	mkdir -p "$TARGET_FOLDER"
+    mkdir -p "$TARGET_FOLDER"
 fi
 
 rm -rf AndroidLibrary
@@ -24,7 +24,7 @@ cat >> AndroidLibrary/AndroidManifest.xml << AXML
 <manifest xmlns:android="http://schemas.android.com/apk/res/android" package="com.unity3d.arcore">
     <uses-feature android:name="android.hardware.camera2" />
     <uses-permission android:name="android.permission.CAMERA" />
-	<uses-sdk android:minSdkVersion="14" android:targetSdkVersion="19" />
+    <uses-sdk android:minSdkVersion="14" android:targetSdkVersion="19" />
 </manifest>
 AXML