+++ package/Source~/source/Providers/InputProvider.h
@@ -9,7 +9,9 @@ public:
 
     virtual void UNITY_INTERFACE_API FillDeviceDefinition(UnityXRInternalInputDeviceId deviceId, IUnityXRInputDeviceDefinition* const deviceDefinition) override;
     virtual bool UNITY_INTERFACE_API UpdateDeviceState(UnityXRInternalInputDeviceId deviceId, IUnityXRInputDeviceState* const deviceState) override;
+    virtual bool UNITY_INTERFACE_API HandleEvent(UnityXRInputEventType eventType, UnityXRInternalInputDeviceId deviceId, void* buffer, unsigned int size);
 
 private:
-    unsigned int m_PoseFeatureIndex;
+    unsigned int m_PositionFeatureIndex;
+    unsigned int m_RotationFeatureIndex;
 };